Fixed broken install script and added antigen

+++ install/dev-tools.sh
@@ -1,44 +1,77 @@
 #!/usr/bin/env bash
-
-read -p "Install zsh+oh-my-zsh+tmux?"           INSTALL_ZSH
-read -p "Install Neovim ppa + neovim pip?"      INSTALL_NEOVIM
-
 set -e
 
+read -p "Install zsh (+antigen) and tmux (+tpm)? "          INSTALL_ZSH
+read -p "Install Neovim ppa + neovim pip? "                 INSTALL_NEOVIM
+
+# Todo: Add antigen
 case $INSTALL_ZSH in
     y|Y)
-        echo "Installing zsh and oh-my-zsh"
+        if [ -z "$ANTIGEN_PATH" ]; then
+            echo "[Error] Setup your dotfiles and source zshrc before attempting to install.."
+            exit 1
+        fi
+
+        echo "Installing zsh"
         sudo apt-get install zsh -y
-        git clone git@github.com:robbyrussell/oh-my-zsh ~/.oh-my-zsh
+
         echo "Installing tmux and tpm"
         sudo apt-get install tmux
-        git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
-        ;;
+        if [ ! -d ~/.tmux/plugins/tpm ]; then
+            echo "Installing tpm"
+            git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+        fi
+
+        if [ ! -f "$ANTIGEN_PATH" ]; then
+            echo "Installing antigen"
+            ANTIGEN_URL=https://raw.githubusercontent.com/zsh-users/antigen/master/antigen.zsh
+            curl -fLo $ANTIGEN_PATH --create-dirs $ANTIGEN_URL
+        fi
+
+        # TODO: Check if zsh is the default shell. If not, ask to set as default.
+        #read -p "Set zsh as the default shell? " DEFAULT_ZSH
+        #case $DEFAULT_ZSH in
+            #y)  chsh -s /bin/zsh
+                #echo "Zsh set as default shell"
+        #esac
 esac
 
 case $INSTALL_NEOVIM in
     y|Y)
+
         # Check so that the env var is set. This probably means our dotfiles are
         # installed. -z checks that the length of the env var is > 0
         if [ -z "$NVIM_DIR" ]; then
-            echo "Setup your dotfiles and source zshrc before attempting to install.."
+            echo "[Error] Setup your dotfiles and source zshrc before attempting to install.."
             exit 1
         fi
 
         if [ ! -d "$NVIM_DIR" ]; then
             echo "...Setting up nvim directories and symlinks..."
             mkdir -p ~/.config/nvim
-            # We can assume dotfiles are set because of $NVIM_DIR being set
-            ln -s ~/.dotfiles/vimrc ~/.config/init.vim
+            # We can assume .dotfiles exist because of $NVIM_DIR being set
+            ln -s ~/.dotfiles/vimrc ~/.config/nvim/init.vim
+        else
+	        echo "Neovim config files already set up"
+        fi
+
+        if [ ! -f "$NVIM_DIR/autoload/plug.vim" ]; then
+            echo "...Installing vim-plug..."
+            PLUG_URL=https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+            curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs $PLUG_URL
+        else
+            echo "...vim-plug already installed..."
         fi
 
         # Neovim from pip needs to be installed for the python plugins to work
-        hash pip 2>/dev/null
+        hash pip || true
         if [ $? -eq 1 ]; then
             sudo wget -O - https://bootstrap.pypa.io/get-pip.py | sudo python
+        else
+	    echo "Pip already installed"
         fi
 
-        hash nvim 2>/dev/null
+        hash nvim || true
         if [ $? -eq 1 ]; then
             echo "...Installing neovim..."
             sudo apt-add-repository ppa:neovim-ppa/unstable -y
@@ -47,17 +80,6 @@ case $INSTALL_NEOVIM in
             sudo pip install neovim
         else
             echo "...Neovim already installed..."
-        fi
-
-        # neovim stores backup files here but that directory isn't created by default
-        #mkdir ~/.local/share/nvim/backup
-        #ln -s ~/.dotfiles/vimrc ~/.config/nvim/init.vim
-        if [ ! -f "$NVIM_DIR/autoload/plug.vim" ]; then
-            echo "...Installing vim-plug..."
-            PLUG_URL=https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-            curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs $PLUG_URL
-        else
-            echo "...vim-plug already installed..."
         fi
 
         ;;
